modularizado/requirements.txt
chromadb
python-dotenv
pypdf
python-docx
groq
fastapi
uvicorn
google-generativeai
google-cloud-speech
